requirements.txt
requests
grpcio
yandexcloud
python-dotenv
yandex
protobuf==3.20.1
googleapis-common-protos==1.53.0
